+++ source/Messaging.Api/MasterDataReceivers/MeteringPointMasterDataResponseListener.cs
@@ -18,6 +18,7 @@ using Energinet.DataHub.MeteringPoints.RequestResponse.Response;
 using Messaging.Api.Configuration.Middleware;
 using Messaging.Application.MasterData;
 using Messaging.Application.Transactions.MoveIn;
+using Messaging.Application.Transactions.MoveIn.MasterDataDelivery;
 using Messaging.Infrastructure.Configuration.InternalCommands;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
