+++ source/Messaging.Application/Transactions/MoveIn/MasterDataDelivery/ForwardCustomerMasterDataHandler.cs
@@ -25,7 +25,7 @@ using Messaging.Domain.Actors;
 using Messaging.Domain.OutgoingMessages;
 using Messaging.Domain.Transactions.MoveIn;
 
-namespace Messaging.Application.Transactions.MoveIn;
+namespace Messaging.Application.Transactions.MoveIn.MasterDataDelivery;
 
 public class ForwardCustomerMasterDataHandler : IRequestHandler<ForwardCustomerMasterData, Unit>
 {
